requirements.txt
pandas==2.1.1
scikit-learn==1.3.1
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
pydantic==2.11.7
numpy==1.26.4
httpx==0.28.1
mlflow==3.1.4
requests==2.32.4
pytest==8.4.1
streamlit==1.47.1
plotly==6.2.0
pydantic-settings==2.10.1
statsmodels==0.14.5
